Start of Python Utility
This starts the creation of the Python version as the BASH version is getting very long and we want to do more complex things.  Given that's the case, BASH is taking more lines of code and longer to develop than performing these same operations in Python.  Especially with the more complex DNS calls and WHOIS lookup processes.